fixed issue where pool was not being passed properly
